Remove the combined proxy to include the Nationale Terminologieserver.
This feature wasn't working anymore, and it never really worked well.
It can be repaired of course, but there are several reasons to deprecate
this feature, at least for now. First and foremost, its imagined use has
never really panned out in practice, and the Nationale
Terminologieserver was never meant for runtime validation anyway, so
this was always just "for testing". At the same time, the options for
validating with the default terminology server have been improved to a
point where it might become bearable.